+++ src/common/commonUI.js
@@ -1,11 +1,16 @@
-define('commonUI',['jquery', 'commonServices'],function($, commonServices){
+define('commonUI',['jquery', 'commonServices', 'jquery.modal'],function($, commonServices){
 
   var commonUI = {
 
     /**
      * Show a modal
      * 
-     *  It allows to use different modal components without affect the engine steps code
+     *  It allows to use different modal components without affect the engine steps code.
+     *
+     *  It uses bootstrap modal as modal library
+     *  
+     *  From version 0.9.30 mbModal (fork of jquery.modal) can be used to show modal. In
+     *  this case, the selector will have the prefix _MBM
      * 
      * == Parameters::
      * 
@@ -15,13 +20,24 @@ define('commonUI',['jquery', 'commonServices'],function($, commonServices){
 
       console.log('show modal');
 
-      // Compatibility mode with jquery.modal plugin
-      if (commonServices.jsBsModalNoConflict && typeof $.fn.bootstrapModal !== 'undefined') {
-        $(selector).bootstrapModal(commonServices.jsBSModalShowOptions());
+      var selectorMBM = selector + '_MBM';
+
+      console.log(selectorMBM);
+      console.log($(selectorMBM).length);
+
+      if ($(selectorMBM).length > 0) {
+        $(selectorMBM).mbModal({ modalClass: "mybooking-modal",
+                                 blockerClass: "mybooking-jquery-modal",});
       }
       else {
-        if ($.fn.modal) {
-          $(selector).modal(commonServices.jsBSModalShowOptions());
+        // Compatibility mode with jquery.modal plugin
+        if (commonServices.jsBsModalNoConflict && typeof $.fn.bootstrapModal !== 'undefined') {
+          $(selector).bootstrapModal(commonServices.jsBSModalShowOptions());
+        }
+        else {
+          if ($.fn.modal) {
+            $(selector).modal(commonServices.jsBSModalShowOptions());
+          }
         }
       }
 
@@ -32,23 +48,34 @@ define('commonUI',['jquery', 'commonServices'],function($, commonServices){
      * 
      *  It allows to use different modal components without affect the engine steps code
      * 
+     *  It uses bootstrap modal as modal library
+     *  
+     *  From version 0.9.30 mbModal (fork of jquery.modal) can be used to show modal. In
+     *  this case, the selector will have the prefix _MBM
+     *
      * == Parameters::
      * 
      * @selector [String] The selector 
-     * 
      * 
      */ 
     hideModal: function(selector) {
 
       console.log('hide modal');
 
-      // Compatibility mode with jquery.modal plugin
-      if (commonServices.jsBsModalNoConflict && typeof $.fn.bootstrapModal !== 'undefined') {
-        $(selector).bootstrapModal('hide');
+      var selectorMBM = selector + '_MBM';
+
+      if ($(selectorMBM).length > 0) {
+        $.mbModal.close();
       }
       else {
-        if ($.fn.modal) {
-          $(selector).modal('hide');
+        // Compatibility mode with jquery.modal plugin
+        if (commonServices.jsBsModalNoConflict && typeof $.fn.bootstrapModal !== 'undefined') {
+          $(selector).bootstrapModal('hide');
+        }
+        else {
+          if ($.fn.modal) {
+            $(selector).modal('hide');
+          }
         }
       }